inital match maker v2 module
Initial creation of the v2 module that is self-containeed. Right now all
types are self-contained and need to be replaced with the real types.
Player update to tags is still needed as a player will only have a
single tag and may even be renamed to "team" in the future.